Found a additional dataset and cleaned some of the data for future use
I have tidied a huge portion of the data currently i want to use the first two datasets in the code separately and the last two as a unit with another additional dataset.

i am also thinking of using one of the first data sets with just the 2018 date to show correlation of it with the third data set in the code.